BackEnd/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
celery==5.3.0
redis==4.5.4
django-redis==5.4.0
tensorflow==2.13.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
nltk==3.8.1
opencv-python==4.8.0
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.3
SpeechRecognition==3.10.0
pydub==0.25.1
python-magic==0.4.27
Pillow==10.0.0 